feat: Add PostgreSQL logging and beta authentication
- Add analysis_logs table schema (db/analysis_logs.sql)
- Implement PostgreSQL connection utilities (utils/postgres_db.py)
- Add beta authentication system (app.py)
- Integrate analysis logging (pages/Analyze_Results.py)
- Add security documentation and deployment guides
- Update requirements.txt with psycopg2-binary
- Add YC feedback implementation
- Enhance UX with DDP/Risk Report style

+++ data_pipeline/fetch_reference_transactions.py
@@ -18,8 +18,8 @@ def process_reference_transactions_data():
         df = pd.read_csv(config.REFERENCE_TRANSACTIONS_FILE)
 
         # Basic data normalization
-        df['fob_price_usd'] = df['fob_price_usd'].astype(float)
-        df['quantity'] = df['quantity'].astype(int)
+        df['fob_price_per_unit'] = df['fob_price_per_unit'].astype(float)
+        df['volume'] = df['volume'].astype(int)
         df['transaction_date'] = pd.to_datetime(df['transaction_date'])
 
         # Save the processed data to a new CSV file
